refactor: moving lens related data structures into `spyglass-lens` crate
