+++ crates/shared/Cargo.toml
@@ -18,4 +18,5 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 strum = "0.24"
 strum_macros = "0.24"
+spyglass-lens = { path = "../spyglass-lens" }
 url = "2.2"
