+++ crates/spyglass-lens/README.md
@@ -1,0 +1,3 @@
+## spyglass-lens
+
+A small library to manage/deal w/ spyglass lens files.
